backend/requirements.txt
torch
torchvision
torchaudio
mediapipe
segment-anything-2
opencv-python
pillow
huggingface_hub
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
numpy
celery
redis
sqlalchemy
alembic
xformers
scipy
scikit-image
pyyaml
weaviate-client
sentence-transformers
google-generativeai
